refactor(wire): extract explicit shape contracts
## Summary

- Move explicit nested structural and serializer validation into the
  1,384-line `_wire_shape_contract.py` subsystem.
- Reduce `_wire.py` while keeping its projection coordinator unchanged.

## Boundaries and compatibility

- Preserve all 26 function bodies and four shape-owned bindings exactly.
- Keep automatic omission of `exclude=True` and `exclude_if` fields
  separate from explicit managed-route rejection.
- Preserve traversal order, diagnostics, and type-only family imports.

## Investigation

- `_wire.py` imports only the shape validator's public internal seam.
- Source and AST comparisons found no missing or changed definitions.
- Three independent reviews found no cycle or compatibility blocker.

## Validation

- `uv run make ci`: passed; 859 tests; 95.88% coverage.
- Python 3.12 with Pydantic 2.12.3: 12 exclusion tests passed.
- Ruff, ty, strict mypy, Vulture, docs, and package build passed.
- Immutable v0.3.0 compatibility contract passed.

Closes #124
Refs #95
